refactor: menu 패키지 생성
