+++ src/main/java/kitchenpos/domain/menu/MenuProduct.java
@@ -1,6 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import kitchenpos.common.BaseDate;
+import kitchenpos.domain.Product;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
