+++ src/main/java/kitchenpos/ui/response/MenuGroupResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.response;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menu.MenuGroup;
 
 import java.util.List;
 import java.util.stream.Collectors;
