+++ src/main/java/kitchenpos/ui/response/MenuProductResponse.java
@@ -1,7 +1,7 @@
 package kitchenpos.ui.response;
 
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuProduct;
 
 import java.util.List;
 import java.util.stream.Collectors;
