+++ src/main/java/kitchenpos/ui/response/MenuResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.response;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.menu.Menu;
 
 import java.math.BigDecimal;
 import java.util.List;
