+++ src/test/java/kitchenpos/domain/MenuTest.java
@@ -1,5 +1,7 @@
 package kitchenpos.domain;
 
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuGroup;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
